[!!!][TASK] Bootstrap: Central constants and path handling
Introduce a class to define main constants and all basic
path evaluations at an early point in bootstrap.

The patch streamlines a lot of the duplicated code for the different
entry points to TYPO3 and moves especially the path handling to a
well documented central location.

Until now, it was documented that is is possible to define own names
for the directories "typo3/", "tslib", "typo3conf/" and "t3lib/".
This never worked without core hacks and was probably never used in
production by anyone for that reason.
This feature is now officially dropped.

Change-Id: Id1ec1082a2572f20500c23357ef5fdbdb05c8498
Resolves: #36889
Releases: 6.0
Reviewed-on: http://review.typo3.org/11030
Reviewed-by: Tolleiv Nietsch
Tested-by: Tolleiv Nietsch
Reviewed-by: Susanne Moog
Tested-by: Susanne Moog
Reviewed-by: Christian Kuhn
Tested-by: Christian Kuhn