[BUGFIX] Update LocalConfiguration fails
Due to changes in function
TYPO3\CMS\Core\Configuration\ConfigurationManager::writeLocalConfiguration
an error is thrown when the configuration file isn't accessable.

Of course when updating the system from a lower version it is not
available. So the installation process has to create the file on its own
to be able to write the settings.

Change-Id: I89553390ebd9f1f576311535e02ec73da6618cd6
Fixes: #42271
Releases: 6.0
Reviewed-on: http://review.typo3.org/16088
Reviewed-by: Christian Kuhn
Tested-by: Christian Kuhn
Reviewed-by: Wouter Wolters
Reviewed-by: Helmut Hummel
Tested-by: Helmut Hummel