[TASK] Remove dummy files from ext:install
The install tool is always bootstrap'ed by typo3/install/index.php,
we can rely on the class loader to find files, so the removed
require_once calls are ok.
The eidHandler for the random key calculation runs through FE
bootstrap, so this is also ok to not require manually.

Change-Id: I45ed37a19039418b1edc440b735809182833ce2d
Resolves: #47848
Related: #47756
Releeases: 6.2
Reviewed-on: https://review.typo3.org/20441
Reviewed-by: Wouter Wolters
Tested-by: Wouter Wolters
Reviewed-by: Anja Leichsenring
Tested-by: Anja Leichsenring