[FEATURE] Add a clear all cache button to install tool
This implements a clear all cache functionality to the install tool.

The implementation is different from the existing backend solution
so that it at least partly clears caches even if some broken
extension is loaded that would kill the operation.

To achieve that, first a low level removal of typo3temp/Cache is
performed, then all table names starting with "cf_" are truncated
with direct database calls. After that ext_localconf and ext_tables
of extension are loaded (this may fatal) and the final cache
configuration is then feeded to the cache manager that executes
the usual flushCaches() call.

Change-Id: Ibd717b0d34bd5b8737a73acebf4632fed38c4d8c
Resolves: #50495
Releases: 6.2
Reviewed-on: https://review.typo3.org/22633
Reviewed-by: Susanne Moog
Tested-by: Susanne Moog
Reviewed-by: Wouter Wolters
Tested-by: Wouter Wolters