[TASK] Install tool configuration system
The patch implements a "configuration system" to the install tool
that bundles several settings to enable them on click
depending on system capabilities.

A "feature" is represented by a set of configuration options that
are filled by different presets. Presets have a priority, the
best preset is the first one. Presets can be available or
not available and active or not. For example the "image handling"
feature has the presets "GraphicsMagick" and "ImageMagick 6" and
a "Custom" preset if the others do not fit.

This first patch implements "Charset conversion", "Image handling",
"Development / Production" and "Extbase object cache" settings.

During installation the available preset with highest priority
are configured.

Change-Id: I7304ef001717989720924e7ae03951eb3874adfd
Resolves: #51454
Releases: 6.2
Reviewed-on: https://review.typo3.org/23321
Reviewed-by: Wouter Wolters
Tested-by: Wouter Wolters
Reviewed-by: Jigal van Hemert
Tested-by: Jigal van Hemert
Reviewed-by: Christian Kuhn
Tested-by: Christian Kuhn