[TASK] Install tool test action fatals if image not exists
The install tool test section creates test images. If permissions
on the target folder are not set up correctly, the file may not
be created, but a correct file location is returned.
In this case not caught exceptions are thrown. This is
now sanitized with a simple file exists view helper to
show some error message instead.

Change-Id: Ic840b0128e5c69f8ce6f601952b8f17e59713e27
Resolves: #51366
Releases: 6.2
Reviewed-on: https://review.typo3.org/23331
Reviewed-by: Simon Schmidt
Tested-by: Simon Schmidt
Reviewed-by: Steffen Ritter
Tested-by: Steffen Ritter