[TASK] Split sys_file table to extract metadata
Parts of the sys_file table are internal cache data
for the FAL Api to work. Other parts are a meta-data
storage for the FAL frontend output.
On the one hand this mixes up several concerns, on
the other hand this makes it difficult to translate
and version metadata since this would result in
duplicate FAL index entries.

As a result those two concerns are split in separate
tables. The user does not "see" the sys_file table
anymore. Metadata is edited only in sys_file_metadata.

Change-Id: Ia6e3664ad5602affcba4e02cf415aa39436d4352
Resolves: #52726
Releases: 6.2
Reviewed-on: https://review.typo3.org/24503
Reviewed-by: Ernesto Baschny
Tested-by: Ernesto Baschny
Reviewed-by: Frans Saris
Tested-by: Frans Saris
Reviewed-by: Fabien Udriot
Tested-by: Fabien Udriot
Reviewed-by: Steffen Ritter
Tested-by: Steffen Ritter