[FEATURE] ext:install Implement template files in FolderStructure
The folder structure definition can be simplified a bit if target
content of files is not stored in the structure definition array
but moved to template files. This also enables handling of binary
files.
The patch implements an option 'targetContentFile' to FileNode and
adapts the current FolderStructure definition to use it.

Change-Id: I097e87a1f7a16e325b398468286fe661abec3553
Resolves: #52351
Releases: 6.2
Reviewed-on: https://review.typo3.org/24122
Reviewed-by: Wouter Wolters
Tested-by: Wouter Wolters
Reviewed-by: Susanne Moog
Tested-by: Susanne Moog