[BUGFIX] ext:install Better error handling in first folder step
During installation it may happen that the folder structure reports
errors during creation, but still creates the main folder structure.
This is the case for example if the document root or fileadmin
is a link and not a directory. In those cases the installer does not
create a LocalConfiguration and shows the first step again. It will
stop with a "ENABLE_INSTALL_TOOL" file missing message afterwards.

This is better sanitized with the patch: If the folder structure
reports errors but still successfully created a typo3conf folder, the
LocalConfiguration is also created and the next step is called after
redirect.

Furthermore, two possible warnings are caught in the folder structure
and adapted status messages are created.

Change-Id: I54d99f826336c513641028f29d389da845d7d26a
Resolves: #52367
Releases: 6.2
Reviewed-on: https://review.typo3.org/24135
Reviewed-by: Anja Leichsenring
Tested-by: Anja Leichsenring
Reviewed-by: Susanne Moog
Tested-by: Susanne Moog
Reviewed-by: Christian Kuhn
Tested-by: Christian Kuhn