[BUGFIX] Base silent updates on LocalConfiguration values only
During silent updates the only file that can be influenced is
LocalConfiguration.php. If values are available in
AdditionalConfiguration.php, those take precedence, but can not be
overridden by the updater. This can lead to endless loops, because
preconditions are never met.

So silent update relies now only on values in LocalConfiguration.
For values in AdditionalConfiguration.php the user must take full
responsibility.

Change-Id: I5b450f0cf09f09bd3bb35a519b5b8d6b33e650ac
Resolves: #52338
Releases: 6.2
Reviewed-on: https://review.typo3.org/24140
Reviewed-by: Christian Kuhn
Tested-by: Christian Kuhn
Reviewed-by: Markus Klein
Reviewed-by: Ernesto Baschny
Tested-by: Ernesto Baschny