[FEATURE] ext:install FolderStructure basic link support
The FolderStructure system currently supports files and directories.
This is now extended with a basic link support that can detect if
nodes are links and point to the specified target if given. The
code is currently fully disabled on Windows and there is also no
support for "fixing" broken structures.

Link nodes are used for index.php, typo3 and typo3_src and are
needed for a future core updater that relies on some structure
conventions like expected core location and link structure.

Change-Id: I5406115b9d44c9a321eb487cbd6ac2f0fab26b1e
Resolves: #52401
Releases: 6.2
Reviewed-on: https://review.typo3.org/24171
Reviewed-by: Wouter Wolters
Tested-by: Wouter Wolters
Reviewed-by: Felix Kopp
Tested-by: Felix Kopp
Reviewed-by: Christian Kuhn
Tested-by: Christian Kuhn