[CLEANUP] Enhance RTE Magic Image FAL migration wizard
If the RTE magic images FAL migration upgrade wizard in the
install tool ever fails (due to max_execution_time) or some
SQL error, then the migration stops and can be run again.

However, if the wizard is run again, it always checks
if the source file still exists. The wizard should rather
check if the target file exists, if not, move the source file
to the target file.

As a separate "step" in this wizard, if the target file (already)
exists, then the DB change can be done.

Additionally, a typo for the output is fixed as well.

Resolves: #53846
Releases: 6.2
Change-Id: Id7c3b6176997848210d83c54cd133819ab15f435
Reviewed-on: https://review.typo3.org/25622
Reviewed-by: Wouter Wolters
Reviewed-by: Tom Ruether
Tested-by: Tom Ruether
Reviewed-by: Ernesto Baschny
Tested-by: Ernesto Baschny