[TASK][CONF] Accept other settings in [SYS][setDBinit]
Since the mysqli interface recommends setting the charset using the
mysqli API the utf8 charset will be initialized when establishing the
database connection using mysqli_set_charset().

Additionally [SYS][setDBinit] can now be set to any value or can totally
be removed. To make sure the database connection still uses the correct
encoding an additional check is added to the database connection process
that checks the MySQL character set session variables.

Finally the old default value of [SYS][setDBinit] will automatically be
removed if it is set to the old default value.

Resolves: #41596
Releases: 6.2
Change-Id: I8d0a9eba50495d52accb59627147c1c87b6a9bb5
Reviewed-on: https://review.typo3.org/15369
Reviewed-by: Wouter Wolters
Reviewed-by: Alexander Opitz
Tested-by: Alexander Opitz
Reviewed-by: Michael Stucki
Tested-by: Michael Stucki
Reviewed-by: Markus Klein
Tested-by: Markus Klein