[TASK] Rename update wizard to upgrade wizard
For the release of TYPO3 4.5 that the wizard in
install tool was renamed to "Upgrade Wizard"
to distinguish between patch level updates and
major version upgrades.

During the refactoring of the install tool for 6.2
the wizard was named "Update wizard again"

Fix the name again and name the upgrade wizard
"Upgrade Wizard" again.

Releases: 6.2
Resolves: #55970
Change-Id: I2cc75b7dfc5ddb3889a080270bf90a101fed935a
Reviewed-on: https://review.typo3.org/27617
Reviewed-by: Markus Klein
Tested-by: Markus Klein
Reviewed-by: Jigal van Hemert
Tested-by: Jigal van Hemert