[TASK] Install Tool Improve Default File/Directory permission warnings
The recommended setup (660/770) should generate a green "ok" and not a
Notice. It now also includes a comment if group read/writeable about
the potential risk and if configured a hint to BE/createGroup.

Also display errors in case the configured mask does not include user
read/writeable flags.

Resolves: #57025
Releases: 6.2
Change-Id: If42d505e0b1bac10cfca1d7b76faec058989ccb2
Reviewed-on: https://review.typo3.org/28492
Reviewed-by: Christian Kuhn
Tested-by: Christian Kuhn
Reviewed-by: Markus Klein
Reviewed-by: Anja Leichsenring
Tested-by: Anja Leichsenring