[BUGFIX] Wrong whitespace handling in RTE-enabled fields upgrade wizard
The upgrade wizards searches for all filepaths in the RTE-enabled fields
as these are found in sys_refindex.ref_string. But in the RTE-enabled
fields spaces are replaced by %20 and in sys_refindex.ref_string not.

This patch replaces whitespaces with %20 in the regularExpression
used the find the occurrences of the filepath.

Resolves: #57143
Releases: 6.2
Change-Id: I6631b476defe7e3faa78914338fce70372ad1fb5
Reviewed-on: https://review.typo3.org/28600
Reviewed-by: Markus Klein
Tested-by: Markus Klein