[BUGFIX] Improve database scheme updater
This patch improves the update process as it removes the hard stop after
an SQL error occurred. All queries should get executed and errors should
be reported at the end. Sometimes it gets necessary that the updater is
called more that once - a notice to the user is added.

Besides the index length is reduced by 1 to ensure the index is set
with limitation in the database and all fields can be updated in the
second step.

Resolves: #57421
Releases: 6.2
Change-Id: I2f830a8a62b8dff3fbe949d40de8ddc30f3dfb9a
Reviewed-on: https://review.typo3.org/28945
Reviewed-by: Jigal van Hemert
Tested-by: Jigal van Hemert
Reviewed-by: Dmitry Dulepov
Tested-by: Dmitry Dulepov
Reviewed-by: Christian Kuhn
Tested-by: Christian Kuhn