[BUGFIX] TceformsUpdateWizard shows up too often
After successfully migrating a field, the field is not added to the list
of finished fields, due to a superfluous check.

Running wizard again finally adds the field to the finished list as the
condition is fulfilled, but the additional run does not
do any useful work anymore.

Fix this by adjusting the condition to safe the extra run.

Resolves: #57497
Releases: 6.2
Change-Id: I6bc15de6f5dbeb58151a1745a849e1b0581b2fbe
Reviewed-on: https://review.typo3.org/29033
Reviewed-by: Frans Saris
Reviewed-by: Alexander Opitz
Tested-by: Alexander Opitz
Reviewed-by: Stefan Neufeind
Tested-by: Stefan Neufeind