[BUGFIX] Use REDIRECT_ prefix as fallback for environment variables
The core uses values of environment variables for
disabling the core updater and to determine the
application context.

When setting the variable with mod_rewrite, certain server environments
prefix the variable name with REDIRECT_.
To ensure compatibility for this case, REDIRECT_
prefix is prepended as a fallback if the regular
variable is not set.

Resolves: #53188
Resolves: #58358
Resolves: #53974
Releases: 6.2
Change-Id: Iaba7f936ff1362d1f70426ec8ad1b1821e4e0d3c
Reviewed-on: https://review.typo3.org/25123
Reviewed-by: Markus Klein
Tested-by: Markus Klein
Reviewed-by: Steffen Müller
Tested-by: Steffen Müller