[TASK] Cache cache_imagesizes using the Caching Framework
Streamline the code by moving the functionality to cache image sizes
processed by GraphicalFunctions (local-only so far) from the database
table cache_imagesizes to the Caching Framework.

The following changes are done:
* use sha1 instead of md5
* use Caching Framework instead of hardcoded DB table
* store only necessary data in the cache

The new Caching Framework configuration is stored
in a new group called "lowlevel" which is only emptied
when pressing the clear cache button in the install tool.

The former table cache_imagesizes has been filled automatically and
never been cleaned except manually in the Install Tool area "Cleanup".
This means there could be a lot of leftover entries. Now the cache is
cleared when using the common clear caches button in the install tool.

Resolves: #28484
Releases: master
Change-Id: Ia68410eb382163e90654718aeb17165dc48e40cc
Reviewed-on: http://review.typo3.org/35686
Reviewed-by: Anja Leichsenring <aleichsenring@ab-softlab.de>
Tested-by: Anja Leichsenring <aleichsenring@ab-softlab.de>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>
Reviewed-by: Stephan Großberndt <stephan@grossberndt.de>
Reviewed-by: Helmut Hummel <helmut.hummel@typo3.org>
Tested-by: Helmut Hummel <helmut.hummel@typo3.org>