[BUGFIX] Fix open_basedir warning when entering install tool
Due to the bug in php https://bugs.php.net/bug.php?id=67378
Warning is triggered when you have open_basedir set up.

This patch suppress the warning as a workaround.
This code is executed only when calling the install tool.
It doesn't influence FE or BE requests so it's acceptable
to use @ operator.

Resolves: #64882
Releases: master
Change-Id: I7881103464ec018cc0a23f47f6cd8c05bb34929c
Reviewed-on: http://review.typo3.org/36680
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>
Reviewed-by: Xavier Perseguers <xavier@typo3.org>
Tested-by: Xavier Perseguers <xavier@typo3.org>