[BUGFIX] Install tool image tests broken
Switching the image manipulation from hardcoded cache table to
the caching framework with #28484 triggers a fatal error in
the install tool image tests. This cache was not properly reset
in the install tool initialization.
The patch resolves this by cleaning up the cache initialization
that has been marked as todo since install tool rewrite. This
is solved for now.
The whole situation is still not perfect since bootstrap, install
tool and extensionmanager manipulate this stuff and the static Cache
class with its static properties in front of the caching framework
does not help to encapsulate and separate the different needs. A
further refactoring releasing the system from static dependencies
and direct global access would help here to reduce complexity with
another patch later.

Resolves: #54498
Related: #28484
Releases: master
Change-Id: Ied01a025ce8e5a3cce03732c95d0914e8fa7af23
Reviewed-on: http://review.typo3.org/37594
Reviewed-by: Nicole Cordes <typo3@cordes.co>
Tested-by: Nicole Cordes <typo3@cordes.co>
Reviewed-by: Markus Klein <klein.t3@reelworx.at>
Tested-by: Markus Klein <klein.t3@reelworx.at>