[!!!][TASK] New Login Screen
The login screen is now supporting background
images and adjustable highlight colors out of the
box. Settings for the login screen can now be
accessed in the backend extension setting.

To keep the focus on the most important stuff,
the news that can be displayed are reduced to
a single view carousel where you can slide
through the latest news.

Releases: master
Resolves: #66431
Change-Id: I12aa4e078cd9782ba4f73921164731a055226004
Reviewed-on: http://review.typo3.org/38366
Reviewed-by: Georg Ringer <georg.ringer@gmail.com>
Tested-by: Georg Ringer <georg.ringer@gmail.com>
Reviewed-by: Wouter Wolters <typo3@wouterwolters.nl>
Tested-by: Wouter Wolters <typo3@wouterwolters.nl>
Reviewed-by: Xavier Perseguers <xavier@typo3.org>
Tested-by: Xavier Perseguers <xavier@typo3.org>
Reviewed-by: Felix Kopp <felix-source@phorax.com>
Tested-by: Felix Kopp <felix-source@phorax.com>
Reviewed-by: Frank Nägler <typo3@naegler.net>
Tested-by: Frank Nägler <typo3@naegler.net>