[TASK] Add support for SPATIAL index type
Add support for "SPATIAL" keys to the SqlSchemaMigrationService.

Ignoring the subpart/key length for this type of index is required as
MySQL does not support specifying the length when creating/modifying
the index but reports this information when retrieving keys for a
table.

This change has been verified against MySQL 5.5.42 (Mac OS X) as well
as Percona Server 5.6.23 (Release 72.1, Linux).

Resolves: #65354
Releases: master
Change-Id: I2a2d0d1e7eddb6ff8fcfb50d88936669e64ea309
Reviewed-on: http://review.typo3.org/38953
Reviewed-by: Stephan Großberndt <stephan@grossberndt.de>
Tested-by: Stephan Großberndt <stephan@grossberndt.de>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>