[!!!][TASK] Remove Cache helper class
The Cache class that acted as a central point when the caching
framework was optionally enabled for 4.3 is now moved
into separate methods.

The duplicate code for creating the database statement is now
moved from the Install Tool to EXT:core to be used in the Extension
Manager as well.

Resolves: #56319
Releases: master
Change-Id: I871a9a6ea506795335e2e8f5b6a410604feba506
Reviewed-on: http://review.typo3.org/40500
Reviewed-by: Nicole Cordes <typo3@cordes.co>
Tested-by: Nicole Cordes <typo3@cordes.co>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>