[TASK] Prevent database error for upgrade wizard
The Upgrade Wizard LanguageIsoCodeUpdate relies on the usage
and therefor existence of ext:static_info_tables. If it is not
available, the upgrade wizard is not needed. It will show up
in the list anyway and cause a database error for the missing
table static_languages.
If no ext:static_info_tables is loaded, don't show the wizard.

Change-Id: I47752efcb2d2994add239ee263969ef44f88df33
Resolves: #68616
Relates: #61542
Releases: master
Reviewed-on: http://review.typo3.org/42087
Reviewed-by: Andreas Fernandez <typo3@scripting-base.de>
Tested-by: Andreas Fernandez <typo3@scripting-base.de>
Reviewed-by: Benjamin Mack <benni@typo3.org>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>