[!!!][TASK] Rename GFX settings referring to IM in their name
Rename GFX settings to more generic names that do not refer to ImageMagick
or no longer supported versions of ImageMagick in their name. The
values for the setting of the graphics processor have been changed to
refer to the tools by name and not by acronym.

The Install Tool has been extended to perform a silent migration of the
affected settings.

Resolves: #43085
Releases: master
Change-Id: I41450eafbadb5782a55f876a87f2bc0ff0704d62
Reviewed-on: https://review.typo3.org/42385
Reviewed-by: Anja Leichsenring <aleichsenring@ab-softlab.de>
Tested-by: Anja Leichsenring <aleichsenring@ab-softlab.de>
Reviewed-by: Georg Ringer <georg.ringer@gmail.com>
Tested-by: Georg Ringer <georg.ringer@gmail.com>