[TASK] Use dedicated database field for `textmedia` media references
To avoid side effects and make maintenance easier the textmedia content
element from EXT:fluid_styled_content uses a dedicated field for media
file references This avoids side effects with EXT:css_styled_content as
the child TCA of the file references is no longer depended on the parent
CType configuration.

This also fixes the wrong media palette being shown in the backend as
the AJAX call to fetch the inline record does not have access to th
parent record.

An upgrade wizard is provided to migrate the file references to the new
assets database field.

Resolves: #70957
Releases: master
Change-Id: I32d7284e2098b4ab74f10f1fc8c7d07a7553511d
Reviewed-on: https://review.typo3.org/44228
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>
Reviewed-by: Anja Leichsenring <aleichsenring@ab-softlab.de>
Tested-by: Anja Leichsenring <aleichsenring@ab-softlab.de>