[TASK] Doctrine: Migrate SqlSchemaMigrationService
The schema migration service has some parts that are ext:dbal
specific which can not be mimik'ed easily at the moment.
To still proceed with the doctrine migration, the original
class is copied over to ext:dbal and registiered as xclass,
while a low-level migration is applied to ext:install
SqlSchemaMigrationService.
This way, the schema migration and load order for now works
with both ext:dbal loaded and not loaded within install tool.
The patch further prepares isolation of ext:dbal and dropping
of a bootstrap database method.

Change-Id: I600229c56160a6eb9f1589c070b5e68a2fc95678
Resolves: #77627
Releases: master
Reviewed-on: https://review.typo3.org/49585
Tested-by: Bamboo TYPO3com <info@typo3.com>
Reviewed-by: Morton Jonuschat <m.jonuschat@mojocode.de>
Tested-by: Morton Jonuschat <m.jonuschat@mojocode.de>
Reviewed-by: Wouter Wolters <typo3@wouterwolters.nl>
Tested-by: Wouter Wolters <typo3@wouterwolters.nl>