[!!!][TASK] Remove TCA l10n_mode=mergeIfNotBlank
The TCA setting "l10n_mode=mergeIfNotBlank" for a single
column is removed from the list of values.

The functionality was there to use the value of a field
of the original record, if the value of the translated record
is empty (or trim'ed empty), and is then overlaid.

The new behaviour is to duplicate the behaviour during
the localize process, and then completely separate.

As a result the related TypoScript setting
config.sys_language_softMergeIfNotBlank is not required
anymore and is removed as well.

Resolves: #79243
Releases: master
Change-Id: I55f3ebd2fe2ddd8412101d5496a0da3c5ab64c68
Reviewed-on: https://review.typo3.org/51239
Tested-by: TYPO3com <no-reply@typo3.com>
Reviewed-by: Susanne Moog <susanne.moog@typo3.org>
Tested-by: Susanne Moog <susanne.moog@typo3.org>
Reviewed-by: Benni Mack <benni@typo3.org>
Tested-by: Benni Mack <benni@typo3.org>