[FEATURE] Add new imageManipulation supporting multiple crop variants
This feature extends the image cropping tool in the backend
so that editors can now not only select one crop area,
but multiple ones per image.

Within the crop are now also a focus are can be selected
and to preview areas that will be covered once the image
is rendered in the frontend one or more cover areas can be configured
to be shown inside the crop area.

This change also adds a format.json view helper and a view helper
to generate backend URIs that are used in the now fully Fluid rendered
imageManipulation element.

This is the TYPO3 integration part. TypeScript and CSS
will be done in a second commit.

Resolves: #75880
Releases: master
Change-Id: I646f0f0a149d05d1f3d8283bcc92ab09aede768e
Reviewed-on: https://review.typo3.org/51515
Tested-by: TYPO3com <no-reply@typo3.com>
Reviewed-by: Frans Saris <franssaris@gmail.com>
Tested-by: Frans Saris <franssaris@gmail.com>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>
Reviewed-by: Andreas Fernandez <typo3@scripting-base.de>
Tested-by: Andreas Fernandez <typo3@scripting-base.de>