[BUGFIX] SchemaMigrator: Don't add indexes on changed columns
Adding or changing an index on MySQL could sometimes fail due
to a dependency between the index definition and the column types.
On MySQL indexes on TEXT/BLOB types might require defining a
prefix length so that the index does not exceed storage limits.

Implement a check when migrating the database in a mode that
only adds fields and indexes such that added indexes do not touch
columns that require changes.

Also update the detection of required database changes in the
upgrade wizard so it does not get stuck on database changes
which will not be performed in that phase.

Resolves: #80096
Resolves: #80043
Releases: master
Change-Id: I4fa403e8429f9f59a6316d0f425d36673a690e31
Reviewed-on: https://review.typo3.org/51928
Reviewed-by: Markus Klein <markus.klein@typo3.org>
Tested-by: TYPO3com <no-reply@typo3.com>
Tested-by: Johannes Kasberger <johannes.kasberger@reelworx.at>
Reviewed-by: Morton Jonuschat <m.jonuschat@mojocode.de>
Tested-by: Morton Jonuschat <m.jonuschat@mojocode.de>