[BUGFIX] Replace .env parsing with reading from environment
Implementing a preliminary parsing code and having an
assumption on the existence of a .env file and a hardcoded
location of that file, makes TYPO3 again implement a speciality
instead of using standard tools.

Therefore the .env reading and parsing is removed in favor
of reading the values directly from the environment, which
gives users the option to choose their own .env solution, or rely
on just setting the env vars.

Resolves: #80626
Releases: master
Change-Id: Id386bb6248e4e814df7579cf848e5b115481b88a
Reviewed-on: https://review.typo3.org/52311
Tested-by: TYPO3com <no-reply@typo3.com>
Reviewed-by: Jan Helke <typo3@helke.de>
Tested-by: Jan Helke <typo3@helke.de>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>
Reviewed-by: Frank Naegler <frank.naegler@typo3.org>
Tested-by: Frank Naegler <frank.naegler@typo3.org>