[TASK] Functionals: Rename "@group mysql" to "@group not-postgres"
Functional tests annotated with "@group mysql" are currently ignored
if running tests on postgres via "--exclude-group mysql" argument.
With mssql we have some tests that run fine with mysql and postgres,
but not with mssql.
To improve handling, "@group mysql" annotation is renamed
to "@group not-postgres". Tests not available on postgres and mssql
will then later have a second annotation "@group not-mssql".
This way its easier to keep track which tests go red on which DBMS.

Change-Id: I2522b00352f1a7d61479588a22cdd9ea5d454ca3
Resolves: #81394
Releases: master, 8.7
Reviewed-on: https://review.typo3.org/52985
Tested-by: TYPO3com <no-reply@typo3.com>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>