[BUGFIX] Install tool: Presets with custom im/gm paths
Fix "Find imagemagick / graphicsmagic in specific direcory" input
field in image preset handling.
The preset handling is still a bit ugly and does not integrate
too well into the ajax based handling all other cards use. This
needs a bigger refactoring and is not done with this patch.
For now, submitting a path to the image preset path input field
reloads the view and executes the path test, so the functionality
is there, it's just a bit ugly to use.

Change-Id: I2da1680b44ff34df05aecc4eb589a293bb8d72c2
Resolves: #82188
Related: #76084
Releases: master
Reviewed-on: https://review.typo3.org/53822
Tested-by: TYPO3com <no-reply@typo3.com>
Reviewed-by: Joerg Boesche <typo3@joergboesche.de>
Tested-by: Joerg Boesche <typo3@joergboesche.de>
Reviewed-by: Susanne Moog <susanne.moog@typo3.org>
Tested-by: Susanne Moog <susanne.moog@typo3.org>