[TASK] Remove suhosin checks
The patch removes suhosin specific checks from ext:install
environment checks. Core v8 needs at least PHP 7.0 and the
suhosin project is still in a pre-alpha "don't use in production"
state and can be considered dead and is of little help in
our application in current state and likely not used anywhere.
For the sake of a cleaner check class, suhosin specific checks
are removed. This has been done in master already with the
related issue.

Change-Id: If003a8db37660662c65290a1e1bf25ef61375b80
Resolves: #82267
Related: #82257
Releases: master, 8.7
Reviewed-on: https://review.typo3.org/53857
Tested-by: TYPO3com <no-reply@typo3.com>
Reviewed-by: Susanne Moog <susanne.moog@typo3.org>
Tested-by: Susanne Moog <susanne.moog@typo3.org>
Reviewed-by: Wouter Wolters <typo3@wouterwolters.nl>
Tested-by: Wouter Wolters <typo3@wouterwolters.nl>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>