[TASK] Move DefaultConfigurationDescription to Yaml format
Defining all configuration options in TYPO3 Install Tool is now handled
via a Yaml file which additionally specifies the "type" of the configuration
option, allowing to use further render settings, one of them already introduced
for "allowedValues". This moves the "All Configuration" to a more flexible output
rendering of dropdowns for a specific type.

Resolves: #82312
Releases: master
Change-Id: Id23c6a8e57177a68659fe98cbe7278f8744c0c34
Reviewed-on: https://review.typo3.org/53923
Tested-by: TYPO3com <no-reply@typo3.com>
Reviewed-by: Mona Muzaffar <mona.muzaffar@gmx.de>
Tested-by: Mona Muzaffar <mona.muzaffar@gmx.de>
Reviewed-by: Stefan Neufeind <typo3.neufeind@speedpartner.de>
Tested-by: Stefan Neufeind <typo3.neufeind@speedpartner.de>
Reviewed-by: Benni Mack <benni@typo3.org>
Tested-by: Benni Mack <benni@typo3.org>