[TASK] Render backend icon list by existing files
Until now all of the backend icons were a manually edited array of file
paths. With this patch defined icon folders in the core are scanned
for existing icon files and build most of that static array
automatically.

For historic reason (and hopefully soon to be removed alltogether) a
second array of statically declared icons is added and merged into.

Some icons that were in the original static list but whose files
actually are missing or misnamed are marked as deprecated.

Resolves: #82348
Releases: master
Change-Id: I8dc4fbbe815d923eda55e1c1b072369e62f66d87
Reviewed-on: https://review.typo3.org/54020
Reviewed-by: Benjamin Kott <benjamin.kott@outlook.com>
Tested-by: Benjamin Kott <benjamin.kott@outlook.com>
Reviewed-by: Andreas Fernandez <typo3@scripting-base.de>
Tested-by: Andreas Fernandez <typo3@scripting-base.de>