[!!!][TASK] Remove tce_db options "prErr" and "uPT"
The two GET/POST parameters "prErr" (Print Errors) and "uPT" (Update Pagetree)
are removed.

The GET/POST parameter "uPT" is there to explicitly request a refresh of
the page tree if visible on the next module load.

The GET/POST parameter "prErr" adds flash messages to the session queue
which should be a default behaviour anyway.

Please note that this only affects the SimpleDataHandlerController (BE module "tce_db") and no datahandler calls itself.

The first parameter of DataHandler->printLogErrorMessages() is removed
as it is unused.

Resolves: #82629
Releases: master
Change-Id: Ifc925e49b49f1537a65141810403a7758e0b448b
Reviewed-on: https://review.typo3.org/54244
Tested-by: TYPO3com <no-reply@typo3.com>
Reviewed-by: Susanne Moog <susanne.moog@typo3.org>
Reviewed-by: Andreas Fernandez <typo3@scripting-base.de>
Tested-by: Susanne Moog <susanne.moog@typo3.org>
Reviewed-by: Frank Naegler <frank.naegler@typo3.org>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>