[TASK] Improved extension configuration API
The patch implements feedback from the new
ExtensionConfiguration API.
get() now falls back to a parsing of ext_conf_template,
so the parser is moved from the install tool
to the ext:core class. This prevents a chicken-egg
issue in early setup phases.

The move of the synchronize* methods forces a raise
of the testing-framework:

composer require --dev typo3/testing-framework ^2.0.3

Change-Id: I9a6700fc66fe78d9df09038a89ea95f8dba81031
Resolves: #83666
Releases: master
Reviewed-on: https://review.typo3.org/55434
Reviewed-by: Helmut Hummel <typo3@helhum.io>
Tested-by: Helmut Hummel <typo3@helhum.io>
Tested-by: TYPO3com <no-reply@typo3.com>
Reviewed-by: Susanne Moog <susanne.moog@typo3.org>
Tested-by: Susanne Moog <susanne.moog@typo3.org>