[FEATURE] New API for the admin panel
The admin panel consisted of one big god class
that contained the complete rendering.

The hook to extend the admin panel only allowed to
add content but not to add new modules (with expandable
headers).

The code has been refactored as a first step for a more
flexible admin panel:

- All modules are now rendered by a class per module
- Modules have an interface
- Modules can be registered in ext_localconf (and overwritten) using
the dependency ordering service for priority
- All new classes are strictly typed

Related: #84044
Resolves: #84045
Releases: master
Change-Id: I124bb503907dcfcbd4425d6f7178b87562d2fda4
Reviewed-on: https://review.typo3.org/55890
Reviewed-by: Frank Naegler <frank.naegler@typo3.org>
Tested-by: TYPO3com <no-reply@typo3.com>
Tested-by: Frank Naegler <frank.naegler@typo3.org>
Reviewed-by: Benni Mack <benni@typo3.org>
Tested-by: Benni Mack <benni@typo3.org>