[BUGFIX] Check registry for upgrade wizard report
In the Reports module the check for unfinished upgrade wizards also
checks the registry to see if a wizard is marked as done. Upgrade
wizards can be marked as done even if their function
shouldRenderWizard() indicates that they are not finished. The
Reports module now shows the same status as the Install Tool.

Resolves: #84146
Releases: 8.7
Change-Id: I994f362545437743216cfc9a64fbf0d6af394379
Reviewed-on: https://review.typo3.org/56049
Tested-by: TYPO3com <no-reply@typo3.com>
Reviewed-by: Andreas Fernandez <a.fernandez@scripting-base.de>
Tested-by: Andreas Fernandez <a.fernandez@scripting-base.de>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>