[FEATURE] Use new REST API for update checks
A new REST api for get.typo3.org was built and should be used
for update checks and additional information gathering.

With this patch the core upgrade and version service is refactored
to use the new API and an additional information in the system
informatian tool bar is displayed.

With the new API it would be possible to add change logs or
links to news for new versions easily in the future.

Additional info:
https://get.typo3.org/v1/api/doc

Resolves: #84549
Releases: master
Change-Id: I9d9c923605a853e9e676367daaa7074bff68aedb
Reviewed-on: https://review.typo3.org/56468
Tested-by: TYPO3com <no-reply@typo3.com>
Reviewed-by: Andreas Fernandez <a.fernandez@scripting-base.de>
Reviewed-by: Alexander Opitz <opitz.alexander@googlemail.com>
Tested-by: Alexander Opitz <opitz.alexander@googlemail.com>
Reviewed-by: Stefan Neufeind <typo3.neufeind@speedpartner.de>
Tested-by: Stefan Neufeind <typo3.neufeind@speedpartner.de>