[BUGFIX] Flush `l10n` cache after downloading languages
This patch let's the Maintenance Area flush the `l10n` caches after
language packs were downloaded. Since the Install Tool uses
"hacked settings", we have to create a new instance of the `CacheManager`
to be able to operate on the mentioned cache. The same handling can be
seen in `ClearCacheService::clearAll()`.

Resolves: #84645
Releases: master
Change-Id: I221bd0d77f2991eac550316586002a8329bb36df
Reviewed-on: https://review.typo3.org/56583
Tested-by: TYPO3com <no-reply@typo3.com>
Reviewed-by: Joshua Westerheide <account@jdoubleu.de>
Tested-by: Joshua Westerheide <account@jdoubleu.de>
Reviewed-by: Anja Leichsenring <aleichsenring@ab-softlab.de>
Tested-by: Anja Leichsenring <aleichsenring@ab-softlab.de>
Reviewed-by: Andreas Fernandez <a.fernandez@scripting-base.de>
Tested-by: Andreas Fernandez <a.fernandez@scripting-base.de>