[TASK] Deprecate Redirecting UrlHandler Hook Concept
The concept for custom UrlHandlers, introduced in TYPO3 v7 for extracting JumpURL
into a TER extension has been deprecated in favor of PSR-15.

These could be registered via
$GLOBALS['TYPO3_CONF_VARS']['SC_OPTIONS']['urlProcessing']['urlHandlers']

The methods related to that functionality have been deprecated.
- $TSFE->initializeRedirectUrlHandlers()
- $TSFE->redirectToExternalUrl()

However, the core-internal calls have been marked with a property to ensure
that no deprecation warning is triggered, but the functionality can be kept.

Resolves: #85124
Releases: master
Change-Id: Ie49d7200618d2b8ffdd5394ea9c649c123719b07
Reviewed-on: https://review.typo3.org/57096
Tested-by: TYPO3com <no-reply@typo3.com>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>
Reviewed-by: Helmut Hummel <typo3@helhum.io>
Tested-by: Helmut Hummel <typo3@helhum.io>